Class system D: hybrid attribute model (locked base + free pool) (#1223) (#1254)
* Assemble class locked base + split the free pool (hybrid attributes) (#1223)

Class system area D (spike #1126): the hybrid attribute model — a level-scaled,
non-reallocatable class locked base plus a reduced free pool.

Backend:
- BattleSnapshot captures ClassId and composes the class's AttributeDistributions
  as a locked base (BaseAmount + AmountPerLevel × level), resolved through a class
  resolver threaded by BattleService into ToBattler/GetModifiers (so it also flows
  into the DefeatRewards power measurement). Fixed composition order — after gear,
  before the proficiency bonuses, ahead of the statics — since float addition isn't
  associative; fail-loud when a captured class can't be resolved.
- GameConstants.StatPointsPerLevel reduced to the free-pool grant (the locked base
  now supplies the bulk of attribute growth).
- NewPlayerFactory seeds the free-pool allocations empty — the locked base provides
  the starting spread, so seeding it would double-count.
- Offline: the locked base is frozen at the snapshot's window-start level like every
  other captured input (the away window is stationary).

Frontend (parity surface, mirroring proficiency #1117):
- Shared class-modifiers.ts computation + bit-exact parity vectors on both sides
  (ClassLockedBaseParityTests / class-locked-base-parity.test.ts), including a
  derived-attribute scenario to pin the additive ordering.
- Regenerated STAT_POINTS_PER_LEVEL.

The live frontend battler wiring + attributes-screen locked/free render ride the
client class-data delivery (#1225), per the proficiency #1117→#1119 precedent.

Closes #1223


* Wire the class locked base into the live frontend battler (#1223)

The e2e suite caught a battle-parity break: every character has a class with a
locked base, so composing it only on the backend (the battle snapshot) made the
frontend's live simulation diverge from the anti-cheat replay — the deferral to
#1225 was wrong (unlike proficiency, the locked base is never empty).

Wire it on both sides now:
- PlayerData carries the player's class attribute distributions
  (LockedBaseDistribution), resolved from the class catalogue in the login
  projection (LoginController.BuildPlayerData).
- PlayerManager memoises classLockedBaseModifiers(distribution, level) and the
  battle engine composes it into the live battler in the same additionalModifiers
  slot as the proficiency bonuses (before the statics, matching the backend order),
  recomputed on level-up since the base is level-scaled. The per-spawn
  change-detection (#811) tracks it by reference.

The class reference delivery (GetClasses) + create-character picker + the
attributes-screen locked/free render stay #1225's scope; the live battler needs
only the player's own fingerprint, delivered with the player.


* Address review: double-arithmetic parity, fail-loud class resolve, tests (#1223)

- Decimal-vs-double parity hole: AttributeDistribution.GetDistributionModifier now
  evaluates BaseAmount + AmountPerLevel × level in double (casting each decimal
  operand first) instead of decimal-then-cast, so it is bit-identical to the
  frontend's IEEE-754 recompute for any authored value — a fractional per-level term
  would otherwise diverge in the last bits and the no-tolerance replay would flag a
  legitimate player. Covers the enemy distribution too (also client-recomputed).
  Added a fractional-per-level parity vector (0.1 × 3 = 0.30000000000000004) to both
  suites; documented in docs/backend.md and the module comments.
- LoginController.BuildPlayerData now fails loud on an unresolvable class (matching
  BattleService.ResolveClass / AccountService), instead of serving an empty
  fingerprint that loads fine but 500s every battle. Reloaded the reference caches in
  the one login fixture (TagsControllerTests) that seeds its player's class directly.
- Battle-engine tests: added the locked-base change-detection case and an assertion
  that reset receives [...lockedBase, ...proficiency] in order with both non-empty.
- Widened the AttributeDistribution contract doc (reused for the class locked base).

Also merged origin/main (proficiency-screen rework) to bring the branch current.


---------

+++ Game.Core/Attributes/AttributeDistribution.cs
@@ -28,16 +28,24 @@ namespace Game.Core.Attributes
         public required decimal AmountPerLevel { get; init; }
 
         /// <summary>
-        /// Creates a new <see cref="AttributeModifier"/> based on the given <paramref name="level"/>.
+        /// Creates a new <see cref="AttributeModifier"/> for the given <paramref name="level"/>:
+        /// <c>BaseAmount + AmountPerLevel × level</c>.
+        /// <para>
+        /// The arithmetic is done in <see cref="double"/> (each authored <see cref="decimal"/> operand is cast
+        /// first), <b>not</b> in decimal-then-cast, so it is bit-identical to the frontend, which recomputes
+        /// the same expression in IEEE-754 double from the JSON-serialized operands (`class-modifiers.ts`,
+        /// `enemy-attributes.ts`). A fractional <see cref="AmountPerLevel"/>/<see cref="BaseAmount"/> would
+        /// otherwise diverge in the last bits between decimal and double arithmetic, and the anti-cheat replay
+        /// compares attributes with no tolerance — so this keeps both the class locked base and the enemy
+        /// distribution exact on the parity surface for any authored value, not just integers.
+        /// </para>
         /// </summary>
-        /// <param name="level"></param>
-        /// <returns></returns>
         public AttributeModifier GetDistributionModifier(int level)
         {
             return new AttributeModifier
             {
                 Attribute = AttributeId,
-                Amount = (double)(BaseAmount + (AmountPerLevel * level)),
+                Amount = (double)BaseAmount + (double)AmountPerLevel * level,
                 Type = EModifierType.Additive,
                 Source = EAttributeModifierSource.AttributeDistribution,
             };
